fix(samba): downgrade to latest stable, no io_uring
The nightly is unsurprisingly unstable. This is a SINK image but based
on Fedora 41, which has Samba 4.21.1.

Because this version of Samba has broken statfs in SMB311 POSIX
extension, this patch disables it, falling back to the legacy UNIX
extensions.

This patch also disables io_uring, which may be causing issues (mount
of specific files failing with -115 in sabnzbd). I have seen one crash
in the io_uring path on the unstable version.